requirements.txt
# PyTorch with CUDA 11.8 support
# Install with: pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118

torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
numpy>=1.20.0
matplotlib>=3.5.0
seaborn>=0.12.0
scikit-learn>=1.0.0
tqdm>=4.65.0
Pillow>=9.0.0
opencv-python>=4.5.0
